Make txtratl a library. Added a separate app and gtests infra.
