+++ CMakeLists.txt
@@ -1,0 +1,9 @@
+cmake_minimum_required(VERSION 3.10)
+
+project("txtratl")
+
+add_subdirectory("txtratl")
+add_subdirectory("tests")
+add_subdirectory("libs/rectpack2d")
+add_subdirectory("libs/stb")
+add_subdirectory("app")
