+++ libs/rectpack2d/LICENSE
