+++ libs/rectpack2d/src/pack.cpp
